requirements.txt
Django==4.2.14
djangorestframework==3.15.2
celery==5.3.6
psycopg2-binary==2.9.9
kombu==5.3.7
amqp==5.2.0